--- api/src/main/java/jakarta/enterprise/lang/model/AnnotationAttribute.java
@@ -1,8 +1,0 @@
-package jakarta.enterprise.lang.model;
-
-// TODO "attribute" is a colloquial expression, perhaps use something closer to the JLS? AnnotationMember?
-public interface AnnotationAttribute {
-    String name();
-
-    AnnotationMember value();
-}
